move schema to config
